move the test cleanup code to the base class
The code no longer does anything backend-specific.
